requirements.txt
PySide6==6.7.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1

